UpstreamIndex: Extract `crates_from_index_head()` method
