+++ src/tests/git.rs
@@ -24,6 +24,29 @@ impl UpstreamIndex {
         Url::from_file_path(&bare()).unwrap()
     }
 
+    /// Obtain a list of crates from the index HEAD
+    pub fn crates_from_index_head(&self, crate_name: &str) -> Vec<cargo_registry::git::Crate> {
+        let repo = &self.repository;
+
+        let path = cargo_registry::git::Repository::relative_index_file(crate_name);
+        let tree = repo.head().unwrap().peel_to_tree().unwrap();
+        let blob = tree
+            .get_path(&path)
+            .unwrap()
+            .to_object(repo)
+            .unwrap()
+            .peel_to_blob()
+            .unwrap();
+        let content = blob.content();
+
+        // The index format consists of one JSON object per line
+        // It is not a JSON array
+        let lines = std::str::from_utf8(content).unwrap().lines();
+        lines
+            .map(|line| serde_json::from_str(line).unwrap())
+            .collect()
+    }
+
     pub fn create_empty_commit(&self) -> anyhow::Result<()> {
         let repo = &self.repository;
